Move util functions from dashboard actions
